pyproject.toml
[project]
name = "autopatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "ansible-runner>=2.4.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "celery[redis]>=5.3.1",
    "docker>=7.1.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langgraph>=1.1.6",
    "minio>=7.2.20",
    "openai>=2.30.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-gvm>=26.11.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "live_llm: requires running vLLM instance with GPU",
    "live_terraform: requires running terraform container and Proxmox access",
]
